Move amountOfBuildingsWithSelectedEdgeMetric #3710
